requirements.txt
Flask
gunicorn
requests
pandas
pyarrow
db-dtypes
google-cloud-storage
google-cloud-bigquery